changed connect structs types

+++ src/structs.h
@@ -1,5 +1,6 @@
 #ifndef STRUCTS_H
 #define STRUCTS_H
+#include <stdint.h>
 
 typedef enum {
     none = 20,
@@ -80,15 +81,15 @@ typedef struct {
 } address_t;
 
 typedef struct {
-    unsigned long long protocol_id; // 0x41727101980 magic constant
-    unsigned long action; // 0 connect
-    unsigned long transaction_id; // random
+    uint64_t protocol_id; // 0x41727101980 magic constant
+    uint32_t action; // 0 connect
+    uint32_t transaction_id; // random
 } connect_request_t;
 
 typedef struct {
-    unsigned long action; // 0 connect
-    unsigned long transaction_id; // same as sent
-    unsigned long long connection_id; // assigned by tracker
+    uint32_t action; // 0 connect
+    uint32_t transaction_id; // same as sent
+    uint64_t connection_id; // assigned by tracker
 } connect_response_t;
 
 #endif //STRUCTS_H
